requirements-setup.txt
# Python dependencies for setup scripts
# Install with: pip install -r requirements-setup.txt

# Required for setup_salesforce_auth.py
requests>=2.31.0
pynacl>=1.5.0
